RasPi/requirements.txt
gpiozero==2.0
paho-mqtt==1.6.1
boto3==1.28.75